Fix browser version
`.type` is only available on elements in cheerio, which caused more than text
nodes to be processed. The solution is to use `.nodeType` instead, which is
available both in cheerio and in the DOM.

Setting `.data` on a text node in the DOM automatically escapes HTML, while in
cheerio it does not. The solution is to change `textNode.data = html` to
`$(textNode).replaceWith(html)`.

I also had to make a slight adjustment to get webpack to set `ENV.browser =
true`.