Make the browser version work without jQuery
This also fixes a few other browser issues:

`.type` is only available on elements in cheerio, which caused more than text
nodes to be processed. The solution is to use `.nodeType` instead, which is
available both in cheerio and in the DOM.

Setting `.data` on a text node in the DOM automatically escapes HTML, while in
cheerio it does not. The solution is to change `textNode.data = html` to
`$(textNode).replaceWith(html)`.

Since browsers do not have the `decodeEntities: false` option, manually escape
text that might be interpreted as HTML otherwise.

For example, if a text node contains the text `a <p> tag` that used to result in
a new paragraph element being created, but doesn't anymore.

I also had to make a slight adjustment to get webpack to set `ENV.browser =
true`.